requirements.txt
aiohttp>=3.11.9
black>=23.12.1
flake8>=6.1.0
isort>=5.13.2
langchain>=0.3.0
langchain-openai>=0.2.11
langgraph>=0.2.56
matplotlib>=3.9.3
numpy>=1.26.4
openai>=1.56.2
bottleneck>=1.3.6
pandas>=2.2.3
pillow>=11.0.0
pydantic>=2.10.3
pytest>=7.4.4
python-dateutil>=2.9.0
python-dotenv>=1.0.0
questionary>=2.1.0
requests>=2.32.3
sqlalchemy>=2.0.36
tabulate>=0.9.0
streamlit>=1.39